# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Endpoints for authentication, login and register\n\n\u003cdetails open\u003e\n\u003csummary\u003e\u003ca href=\"javascript:void(0)\" style=\"cursor: pointer !important;\"\u003e📋 Message Codes\u003c/a\u003e\u003c/summary\u003e\n\n| Code   | Description                   |\n| ------ | ------------------------------|\n| 1000   | successful                    |\n| 1001   | unsuccessful                  |\n| 1002   | found                         |\n| 1003   | not found                     |\n| 4000   | internal error                |\n| 4001   | invalid request body          |\n| 4002   | invalid authentication token  |\n| 4003   | missing authentication token  |\n| 4004   | invalid signature             |\n| 4005   | missing signature             |\n| 4006   | invalid appid                 |\n| 4007   | missing appid                 |\n| 4008   | invalid query parameters      |\n\n\u003c/details\u003e",
        "title": "LBE API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:18080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/gr": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Checks if a GR member ID is already registered.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Verify GR member existence",
                "parameters": [
                    {
                        "description": "GR registration check payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RegisterGr"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GR member found",
                        "schema": {
                            "$ref": "#/definitions/responses.GrExistenceSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON request body",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized – API key missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/gr-cms": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Checks if a GR CMS member email is in the system and caches their profile for follow‑up.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Verify and cache GR CMS member",
                "parameters": [
                    {
                        "description": "GR CMS register payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RegisterGrCms"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email not found; profile cached",
                        "schema": {
                            "$ref": "#/definitions/responses.GrCmsExistenceSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON request body",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized – API key missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email already registered",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/gr-reg/{reg_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a temporarily cached GR CMS profile by registration ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get cached GR CMS profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Registration ID",
                        "name": "reg_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cached profile found",
                        "schema": {
                            "$ref": "#/definitions/responses.CachedGrCmsSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Registration ID is required",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Cached profile not found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Registers a new user record in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create new user",
                "parameters": [
                    {
                        "description": "User create payload",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/responses.CreateSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON request body",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized – API key missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/register/verify": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Checks if an email is already registered; if not, sends an OTP for signup.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Verify email for registration",
                "parameters": [
                    {
                        "description": "Registration request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.VerifyUserExistence"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email not registered; OTP sent",
                        "schema": {
                            "$ref": "#/definitions/responses.RegisterSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON request body",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized – API key missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email already registered",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.GrMember": {
            "type": "object",
            "properties": {
                "dob": {
                    "description": "Dob is the date of birth in YYYY-MM-DD format.",
                    "type": "string",
                    "example": "1985-04-12"
                },
                "email": {
                    "description": "Email is the member's email address.",
                    "type": "string",
                    "example": "jane.doe@example.com"
                },
                "f_name": {
                    "description": "FirstName is the given name of the GR member.",
                    "type": "string",
                    "example": "Jane"
                },
                "gr_id": {
                    "description": "GrId is the unique GR member identifier.",
                    "type": "string",
                    "example": "GR12345"
                },
                "l_name": {
                    "description": "LastName is the family name of the GR member.",
                    "type": "string",
                    "example": "Doe"
                },
                "mobile": {
                    "description": "Mobile is the contact phone number.",
                    "type": "string",
                    "example": "98765432"
                }
            }
        },
        "model.Otp": {
            "type": "object",
            "properties": {
                "otp": {
                    "description": "Otp is the one‑time password sent to the user.\nexample: \"123456\"",
                    "type": "string",
                    "example": "123456"
                },
                "otp_expiry": {
                    "description": "OtpExpiry is the Unix timestamp (seconds since epoch) when the OTP expires.\nexample: 1744176000",
                    "type": "integer",
                    "example": 1744176000
                }
            }
        },
...
        "requests.Referral": {
            "type": "object",
            "properties": {
                "referral_code": {
                    "description": "Code of the referring customer",
                    "type": "string",
                    "example": "JOHN-70A756"
                }
            }
        },
        "requests.RegisterGr": {
            "type": "object",
            "required": [
                "gr_id",
                "gr_pin"
            ],
            "properties": {
                "gr_id": {
                    "description": "GR system identifier for the member.",
                    "type": "string",
                    "example": "GR12345"
                },
                "gr_pin": {
                    "description": "PIN code associated with the GR member.",
                    "type": "string",
                    "example": "9876"
                }
            }
        },
        "requests.RegisterGrCms": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "dob": {
                    "description": "Dob is the date of birth in YYYY-MM-DD format.",
                    "type": "string",
                    "example": "1985-04-12"
                },
                "email": {
                    "description": "Email is the member's email address.",
                    "type": "string",
                    "example": "jane.doe@example.com"
                },
                "f_name": {
                    "description": "FirstName is the given name of the GR member.",
                    "type": "string",
                    "example": "Jane"
                },
                "gr_id": {
                    "description": "GrId is the unique GR member identifier.",
                    "type": "string",
                    "example": "GR12345"
                },
                "l_name": {
                    "description": "LastName is the family name of the GR member.",
                    "type": "string",
                    "example": "Doe"
                },
                "mobile": {
                    "description": "Mobile is the contact phone number.",
                    "type": "string",
                    "example": "98765432"
                },
                "url": {
                    "description": "URL to send the registration confirmation link to.",
                    "type": "string",
                    "example": "https://example.com/confirm?reg_id=abc123"
                }
            }
        },
        "requests.RegisterUser": {
            "type": "object",
            "properties": {
                "sign_up_type": {
                    "type": "string",
                    "example": "NEW"
                },
                "users": {
                    "$ref": "#/definitions/requests.UserRequest"
                }
            }
        },
...
        "requests.UserRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address line 1",
                    "type": "string",
                    "example": "7 Tremont Street"
                },
                "address2": {
                    "description": "Address line 2",
                    "type": "string",
                    "example": "8 Tremont Street"
                },
                "city": {
                    "description": "City of residence",
                    "type": "string",
                    "example": "Boston"
                },
                "country": {
                    "description": "3-letter ISO-3166-1 country code",
                    "type": "string",
                    "example": "USA"
                },
                "dob": {
                    "description": "Date of birth, YYYY-MM-DD",
                    "type": "string",
                    "example": "1980-01-01"
                },
                "email": {
                    "description": "Customer’s email (encrypted at rest). Required if ExternalID is not specified.",
                    "type": "string",
                    "example": "john.smith@test.com"
                },
                "external_id": {
                    "description": "Identifier for customer in external system.\nRequired if Email is not specified.",
                    "type": "string",
                    "example": "1284111"
                },
                "external_id_type": {
                    "description": "Type associated with external identifier.",
                    "type": "string",
                    "example": "facebook"
                },
                "first_name": {
                    "description": "First name",
                    "type": "string",
                    "example": "John"
                },
                "gender": {
                    "description": "Gender of customer: m or f",
                    "type": "string",
                    "example": "m"
                },
                "ip": {
                    "description": "Customer’s IP at registration.",
                    "type": "string",
                    "example": "203.0.113.42"
                },
                "last_name": {
                    "description": "Last name (surname)",
                    "type": "string",
                    "example": "Smith"
                },
                "locale": {
                    "description": "Preferred locale, e.g. en-US",
                    "type": "string",
                    "example": "en-US"
                },
                "opted_in": {
                    "description": "Indicates opt-in to loyalty program (defaults to true if not set).",
                    "type": "boolean",
                    "example": true
                },
                "phone_numbers": {
                    "description": "Phone numbers associated with customer",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.PhoneNumber"
                    }
                },
                "referral": {
                    "description": "Referral info when signing up with a referrer code",
                    "allOf": [
                        {
                            "$ref": "#/definitions/requests.Referral"
                        }
                    ]
                },
                "state": {
                    "description": "State/province/region",
                    "type": "string",
                    "example": "MA"
                },
                "zip": {
                    "description": "Zip or postal code",
                    "type": "string",
                    "example": "02021"
                }
            }
        },
        "requests.VerifyUserExistence": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "Email address to check for existing registration.",
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
...
        "responses.CachedGrCmsSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "in: body",
                    "type": "integer",
                    "example": 1002
                },
                "data": {
                    "$ref": "#/definitions/model.GrMember"
                },
                "message": {
                    "type": "string",
                    "example": "cached profile found"
                }
            }
        },
        "responses.CreateSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "in: body",
                    "type": "integer",
                    "example": 1000
                },
                "data": {
                    "$ref": "#/definitions/responses.User"
                },
                "message": {
                    "type": "string",
                    "example": "user created"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is your internal API status code, e.g. 1002",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "description": "Message is a human‑readable description, e.g. \"invalid json request body\"",
                    "type": "string"
                }
            }
        },
...
        "responses.GrCmsExistenceSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "in: body",
                    "type": "integer",
                    "example": 1003
                },
                "data": {
                    "$ref": "#/definitions/model.GrMember"
                },
                "message": {
                    "type": "string",
                    "example": "email not found"
                }
            }
        },
        "responses.GrExistenceSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "in: body",
                    "type": "integer",
                    "example": 1000
                },
                "data": {
                    "$ref": "#/definitions/model.GrMember"
                },
                "message": {
                    "type": "string",
                    "example": "successful"
                }
            }
        },
...
        "responses.RegisterSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "in: body",
                    "type": "integer",
                    "example": 1000
                },
                "data": {
                    "$ref": "#/definitions/model.Otp"
                },
                "message": {
                    "type": "string",
                    "example": "email not found"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer \u003cyour-jwt\u003e\" to authorize",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}